+++ szalkszop/szalkszop/Areas/Admin/Controllers/PaymentMethodController.cs
@@ -18,11 +18,40 @@ namespace szalkszop.Areas.Admin.Controllers
 		public ActionResult Index()
 		{
 			var viewModel = _paymentMethodService.GetList();
+
 			return View(viewModel);
+		}
+
+		public ActionResult Create()
+		{
+			var viewModel = new PaymentMethodViewModel
+			{
+				Heading = "Create",
+			};
+
+			return View("PaymentMethodForm", viewModel);
+		}
+
+		[HttpPost]
+		public ActionResult Create([Bind(Exclude = "Id")]PaymentMethodViewModel viewModel)
+		{
+			if (!ModelState.IsValid)
+			{
+				viewModel.Heading = "Create";
+
+				return View("PaymentMethodForm", viewModel);
+			}
+
+			_paymentMethodService.AddPaymentMethod(viewModel);
+
+			return RedirectToAction("Index");
 		}
 
 		public ActionResult Edit(int id)
 		{
+			if (!_paymentMethodService.DoesPaymentMethodExist(id))
+				return HttpNotFound();
+
 			var paymentMethod = _paymentMethodService.GetPaymentMethod(id);
 
 			var viewModel = new PaymentMethodViewModel
@@ -39,6 +68,15 @@ namespace szalkszop.Areas.Admin.Controllers
 		[HttpPost]
 		public ActionResult Edit(PaymentMethodViewModel viewModel)
 		{
+			if (!_paymentMethodService.DoesPaymentMethodExist(viewModel.Id))
+				return HttpNotFound();
+
+			if (!ModelState.IsValid)
+			{
+				viewModel.Heading = "Edit";
+				return View("PaymentMethodForm", viewModel);
+			}
+
 			_paymentMethodService.Edit(viewModel);
 
 			return RedirectToAction("Index");
@@ -46,23 +84,10 @@ namespace szalkszop.Areas.Admin.Controllers
 
 		public ActionResult Delete(int id)
 		{
+			if (!_paymentMethodService.DoesPaymentMethodExist(id))
+				return HttpNotFound();
+
 			_paymentMethodService.DeletePaymentMethod(id);
-
-			return RedirectToAction("Index");
-		}
-
-		public ActionResult Create()
-		{
-			var viewModel = new PaymentMethodViewModel();
-			viewModel.Heading = "Create";
-
-			return View("PaymentMethodForm", viewModel);
-		}
-
-		[HttpPost]
-		public ActionResult Create(PaymentMethodViewModel viewModel)
-		{
-			_paymentMethodService.AddPaymentMethod(viewModel);
 
 			return RedirectToAction("Index");
 		}
